moved permission resposibility

+++ API/record/delete.php
@@ -18,6 +18,7 @@ include_once $GLOBALS["webroot"] . "/core/core.php";
     **************************
 */
 
+$token = null;
 if(!isset($_POST["token"])){
     session_start();
     if(!isset($_SESSION["username"])){
@@ -25,7 +26,9 @@ if(!isset($_POST["token"])){
         echo "<h1>No token!</h1>";
         exit(0);
     }
-} 
+}else{
+    $token=$_POST["token"];
+}
 
 if(!isset($_POST["record"])){
     header("HTTP/1.1 400 Bad request");
@@ -47,6 +50,13 @@ if(!isset($_POST["value"])){
 
 include_once $GLOBALS["webroot"] . "/core/permissions.php";
 include_once $GLOBALS["webroot"] . "/core/manageDomainDatabase.php";
+
+$domainPermission = $_POST["record"] . ".delete";
+if(!userHasAnyOfThesePermissions(array($domainPermission, "admin"))){
+    header("HTTP/1.1 403 Forbidden");
+    echo "<h1>;|!</h1>";
+    exit(0);
+}
 
 deleteRecord($_POST["record"], $_POST["type"], $_POST["value"]);
